Add files via upload
GroundingDINO (MMDetection) was used as the base model, with installation instructions provided in the documentation. The model was trained on a quadrant-enumerated DENTEX subset, extended with six dental abnormality annotations. DENTEX repository provides the original X-ray images. Acknowledgements are given to the developers of GroundingDINO and MMDetection, and the dentists who annotated the data.